+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 1.0.5 -> 1.0.6
  * Updated bindings and upgraded to latest Chartboost SDKs (iOS 6.5.0, Android 6.5.1)
+ * Removed iOS binding for closeImpression, as it was removed from the iOS SDK.
  * Added bindings for getAutoCacheAds, setAutoCacheAds.
 
 ## 1.0.4 -> 1.0.5